fix: importing useSearchPrams didn't patch history API
