+++ src/useLocation.ts
@@ -1,27 +1,6 @@
 import { useEffect, useState } from 'react';
 import { isBrowser, off, on } from './misc/util';
-
-const patchHistoryMethod = (method) => {
-  const history = window.history;
-  const original = history[method];
-
-  history[method] = function (state) {
-    const result = original.apply(this, arguments);
-    const event = new Event(method.toLowerCase());
-
-    (event as any).state = state;
-
-    window.dispatchEvent(event);
-
-    return result;
-  };
-};
-
-if (isBrowser) {
-  patchHistoryMethod('pushState');
-  patchHistoryMethod('replaceState');
-}
-
+import tryPatchHistoryMethod from './misc/patchHistoryMethod';
 export interface LocationSensorState {
   trigger: string;
   state?: any;
@@ -36,6 +15,8 @@ export interface LocationSensorState {
   protocol?: string;
   search?: string;
 }
+
+tryPatchHistoryMethod();
 
 const useLocationServer = (): LocationSensorState => ({
   trigger: 'load',
